Dev #117: initial refactoring into a CLI stub, investigating further dissolving of code
