--- cats/__main__.py
@@ -1,4 +1,0 @@
-from . import main
-
-if __name__ == "__main__":
-    main()
